max_age=1 test

See merge request openid/conformance-suite!864

+++ src/main/java/net/openid/conformance/condition/client/CheckIdTokenAuthTimeClaimsSameIfPresent.java
@@ -21,8 +21,8 @@ public class CheckIdTokenAuthTimeClaimsSameIfPresent extends AbstractCondition {
 		JsonObject secondIdToken = env.getObject("id_token").getAsJsonObject("claims");
 
 		if (firstIdToken.has(CLAIM_AUTH_TIME) && secondIdToken.has(CLAIM_AUTH_TIME)) {
-			var firstAuthTime = OIDFJSON.getInt(firstIdToken.get(CLAIM_AUTH_TIME));
-			var authTime = OIDFJSON.getInt(secondIdToken.get(CLAIM_AUTH_TIME));
+			var firstAuthTime = OIDFJSON.getLong(firstIdToken.get(CLAIM_AUTH_TIME));
+			var authTime = OIDFJSON.getLong(secondIdToken.get(CLAIM_AUTH_TIME));
 
 			if (firstAuthTime != authTime) {
 				throw error("The id_tokens contain different auth_time claims, but must contain the same auth_time.", args("first_id_token", firstIdToken, "second_id_token", secondIdToken));
